Refactored databoundary to datamanager
