+++ application/CMakeLists.txt
@@ -23,7 +23,7 @@ add_executable (
 target_link_libraries (
     console_app
     auth
-    databoundary
+    logindata
     utility
     ${STATIC_LIBS}
 )
