+++ datamanager/CMakeLists.txt
@@ -1,0 +1,4 @@
+project (datamanager)
+
+add_subdirectory (logindata)
+add_subdirectory (storage)
