+++ datamanager/storage/CMakeLists.txt
@@ -1,0 +1,16 @@
+project (storage)
+
+# Databoundary lib
+add_library (
+    dummydata
+    STATIC
+    # in-memory db
+    stackdb.hpp
+    stackdb.cpp
+)
+
+
+target_link_libraries (
+    dummydata
+    entity
+)
